// app/types/newItemProps.ts
interface itemProps {
  etag: string;
  id: {
    kind: string;
    videoId: string;
  };
  kind: string;
  snippet: {
    channelId: string;
    channelTitle: string;
    description: string;
    liveBroadcastContent: string;
    publishTime: string;
    publishedAt: string;
    thumbnails: {
      default: {
        url: string;
        width: number;
        height: number;
      };
      high: { url: string; width: number; height: number };
      medium: { url: string; width: number; height: number };
    };
    title: string;
  };
}

export type { itemProps };